requirements.txt
Django==2.2.5
Pillow==6.2.0
django-crispy-forms
django-phonenumber-field
phonenumbers
